@@ -1,13 +1,14 @@
 {
   "name": "react-image-video-lightbox",
-  "version": "2.0.4",
+  "version": "3.0.0-rc1",
   "description": "A React lightbox that supports videos, images and pinch zooming on images. Optimized for mobile UI with swiping, but can be used on desktop as well.",
-  "main": "build/index.js",
+  "main": "lib/index.js",
+  "module": "es/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack --watch",
-    "build": "webpack",
-    "prepublish": "npm run build"
+    "build": "nwb build-react-component",
+    "clean": "nwb clean-module && nwb clean-demo",
+    "prepublishOnly": "npm run build",
+    "start": "nwb serve-react-demo"
   },
   "keywords": [
     "lightbox",
@@ -25,23 +26,17 @@
   "repository": "https://github.com/Ngineer101/react-image-video-lightbox",
   "bugs": "https://github.com/Ngineer101/react-image-video-lightbox/issues",
   "files": [
-    "build/index.js"
+    "css",
+    "es",
+    "lib",
+    "umd"
   ],
-  "dependencies": {
-    "@material-ui/core": "1.3.1",
-    "@material-ui/icons": "1.1.0",
-    "react": "16.4.1",
-    "react-dom": "16.4.1"
+  "peerDependencies": {
+    "react": "16.x"
   },
   "devDependencies": {
-    "babel-cli": "6.26.0",
-    "babel-core": "6.26.3",
-    "babel-loader": "7.1.4",
-    "babel-plugin-transform-object-rest-spread": "6.26.0",
-    "babel-plugin-transform-react-jsx": "6.24.1",
-    "babel-preset-env": "1.7.0",
-    "url-loader": "1.0.1",
-    "webpack": "4.12.2",
-    "webpack-command": "0.3.1"
+    "nwb": "0.25.x",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   }
 }
